compute rgb values in cuda kernels

+++ cuda-fractals/include/fractal_cuda.h
@@ -7,5 +7,5 @@
 #ifdef CUDA
 extern bool fractal_cuda_init(int width, int height);
 extern void fractal_cuda_clean();
-extern void fractal_cuda_get_colors(int * image, struct FractalProperties * fp);
+extern void fractal_cuda_get_colors(uint8_t * image, struct FractalProperties * fp);
 #endif
